Rename varname/varname.py to varname/core.py
