+++ varname/__init__.py
@@ -1,6 +1,6 @@
 """Dark magics about variable names in python"""
 # pylint: disable=unused-import
 from .utils import VarnameRetrievingError, config
-from .varname import varname, nameof, will
+from .core import varname, nameof, will
 
 __version__ = "0.6.0"
